+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSFE9BPGM3XCE43G43STH5DM"
+        "01KSFTNWCMGTWB5J939HCB0R3N"
     )
     r"""The unique identifier of the data source"""