+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -113,7 +113,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-25T11:28:11.905Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-25T15:04:44.027Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")