+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSFE9BPFVZK1F5AV83M6CQVN"
+        "01KSFTNWCKDACGCTS9PG8PEADX"
     )
     r"""The unique identifier of the data source"""