+++ packages/orq-rc/RELEASES.md
@@ -8169,3 +8169,13 @@ Based on:
 - [python v4.10.0-rc.31] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.31] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.31 - packages/orq-rc
+
+## 2026-05-25 15:07:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.765.0 (2.886.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.32] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.32] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.32 - packages/orq-rc